requirements.txt
pycryptodome
passlib
pyotp
bcrypt
psycopg2_binary
flup
python-ldap
pymysql
qrcode